requirements.txt
asgiref==3.7.2
Django==3.2.19
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
pymongo==3.12.1
pytz==2023.3
sqlparse==0.2.4
typing-extensions==4.6.3
